requirements_rag.txt
flask==3.0.0
google-generativeai==0.3.2
chromadb==0.4.22
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
